Revert "Improve mcp (#37)"
This reverts commit 5ae3645fd494880feb42a0ebd77250812044a745.
