+++ smithery.yaml
@@ -25,4 +25,4 @@ build:
   dockerBuildPath: "."
 
 env:
-  NODE_ENV: "production"
+  NODE_ENV: "production"
