+++ src/app.tsx
@@ -7,11 +7,12 @@ import {
 import Layout from "@/components/layout";
 import InspectorPage from "@/pages/inspector-page";
 import ChatPage from "@/pages/chat-page";
-import { AuthProvider } from "@/lib/auth-context";
-import { ApiKeysProvider } from "@/lib/api-keys-context";
+import { AuthProvider, useAuth } from "@/lib/auth-context";
 // import TurnstileAuth from "@/components/turnstile-auth";
 
 function AppContent() {
+  const { isAuthenticated } = useAuth();
+
   return (
     <>
       <Routes>
@@ -28,13 +29,11 @@ function AppContent() {
 export default function App() {
   return (
     <AuthProvider>
-      <ApiKeysProvider>
-        <Router>
-          <Layout>
-            <AppContent />
-          </Layout>
-        </Router>
-      </ApiKeysProvider>
+      <Router>
+        <Layout>
+          <AppContent />
+        </Layout>
+      </Router>
     </AuthProvider>
   );
 }
